refactor() replace activities
