+++ app/src/main/java/com/example/taskraken/activities/RegisterActivity.java
@@ -1,9 +1,8 @@
-package com.example.taskraken;
+package com.example.taskraken.activities;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Intent;
-import android.content.res.Resources;
 import android.os.Bundle;
 import android.widget.Button;
 import android.widget.EditText;
@@ -13,12 +12,10 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import com.example.taskraken.R;
 import com.example.taskraken.network.api.AuthApi;
 import com.example.taskraken.network.schemas.users.UserCreate;
 import com.example.taskraken.network.services.NetworkService;
-
-import org.json.JSONException;
-import org.json.JSONObject;
 
 import java.io.IOException;
